pyproject.toml
[project]
name = "my-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.134.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "python-multipart>=0.0.9",
    "pydantic[email]==2.7.0"
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pytest>=9.0.2",
    "ruff>=0.15.4",
]
